move generalized kernel from experimental to limbo
