+++ src/tests/test_gp.cpp
@@ -29,7 +29,11 @@ Eigen::VectorXd make_v2(double x1, double x2)
 }
 
 struct Params {
-    struct kernel_maternfivehalfs {
+    struct SquaredExpARD : public defaults::SquaredExpARD {
+  };
+
+
+  struct kernel_maternfivehalfs {
         BO_PARAM(double, sigma, 1);
         BO_PARAM(double, l, 0.25);
     };
