+++ src/tests/test_init_functions.cpp
@@ -30,7 +30,10 @@ struct Params {
         BO_PARAM(int, iterations, 0);
     };
 
-    struct kernel_maternfivehalfs {
+  struct SquaredExpARD : public defaults::SquaredExpARD {
+  };
+
+  struct kernel_maternfivehalfs {
         BO_PARAM(double, sigma, 1);
         BO_PARAM(double, l, 0.25);
     };
